Extract risk drawdown response module
